feat: clean up routes
