+++ .env.example
@@ -1,7 +1,1 @@
 OPENAI_API_KEY=
-OXFORD_CHITCHAT_APP_ID=
-OXFORD_APP_KEY=
-PGHOST=
-PGDATABASE=
-PGUSER=
-PGPASSWORD=
