+++ requirements.txt
@@ -1,35 +1,69 @@
+aiofiles==23.2.1
 aiohttp==3.8.3
 aiosignal==1.3.1
+amqp==5.1.1
 asgiref==3.6.0
 async-timeout==4.0.2
 attrs==22.2.0
-blinker==1.6.2
+billiard==4.1.0
+blinker==1.5
+celery==5.3.4
 certifi==2022.12.7
 charset-normalizer==2.1.1
 click==8.1.7
+click-didyoumean==0.3.0
+click-plugins==1.1.1
+click-repl==0.3.0
+colorama==0.4.6
+cors==1.0.1
 Django==4.1.5
 django-cors-headers==3.13.0
 djangorestframework==3.14.0
 dnspython==2.4.2
+filelock==3.12.4
 Flask==2.3.3
 Flask-Cors==4.0.0
 Flask-PyMongo==2.3.0
 frozenlist==1.3.3
+future==0.18.3
+gevent==23.9.1
+greenlet==3.0.0rc3
 gunicorn==20.1.0
+h11==0.14.0
+h2==4.1.0
+hpack==4.0.0
+hypercorn==0.14.4
+hyperframe==6.0.1
 idna==3.4
 itsdangerous==2.1.2
 Jinja2==3.1.2
+kombu==5.3.2
 MarkupSafe==2.1.3
 multidict==6.0.4
+nanoid==2.0.0
 openai==0.28.0
+priority==2.0.0
+prompt-toolkit==3.0.39
 psycopg2==2.9.5
 pymongo==4.5.0
+PySocks==1.7.1
+python-dateutil==2.8.2
 python-decouple==3.7
 python-dotenv==1.0.0
 pytz==2022.7
+quart==0.18.4
 requests==2.28.2
+requests-file==1.5.1
+six==1.16.0
 sqlparse==0.4.3
+tldextract==3.5.0
 tqdm==4.64.1
+tzdata==2023.3
 urllib3==1.26.14
+vine==5.0.0
+wcwidth==0.2.6
 Werkzeug==2.3.7
+wsproto==1.2.0
 yarl==1.8.2
+zope.event==5.0
+zope.interface==6.0
